Add files via upload
We will set up a cron job to run every 6 hours in order to check the status of orders. Based on the order status and pending we will send an email to the customer informing them that our team will check the order and get back to them as soon as possible. Additionally, another email will be sent to the admin notifying them about the type of order received and requesting them to check it at their earliest convenience.